Create README.md
This project analyzes Uber and Lyft prices to help users know which option is cheapest, with respect to the day, time, and location of pick-up / drop-off. The dataset has nearly 700,000 rows of data, taken from less than 4 weeks (Nov. 25 - Dec. 18) in the Boston area.  A python Jupiter notebook was used and all the packages used are imported in the notebook. 

*Note: not all of the figures used in this report are in the figures section.